refactor: mv ft_pretty_print into test_hashtable
