+++ tests/test_hashtable.c
@@ -6,11 +6,48 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 13:29:52 by gwolf             #+#    #+#             */
-/*   Updated: 2023/06/06 16:11:54 by gwolf            ###   ########.fr       */
+/*   Updated: 2023/07/30 14:55:10 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "test.h"
+
+/**
+ * @brief Pretty print all elements.
+ *
+ * Print starting string.
+ * Print all elements at each index. Every element is denoted by a "->".
+ * If more than one element is at a given index they get separated by newline.
+ * Print number of elements.
+ * Print end string.
+ * @param ht Hashtable to print.
+ */
+void	ft_hashtable_pretty_print(t_hashtable *ht)
+{
+	uint32_t	i;
+	t_env_var	*tmp;
+
+	printf("Start Table\n");
+	i = 0;
+	while (i < ht->size)
+	{
+		if (ht->elements[i] == NULL)
+			printf("%i\n---\n", i);
+		else
+		{
+			printf("%i\n", i);
+			tmp = ht->elements[i];
+			while (tmp != NULL)
+			{
+				printf("->\t%s\n", tmp->env_string);
+				tmp = tmp->next;
+			}
+		}
+		i++;
+	}
+	printf("Number of elements: %d\n", ht->num_elements);
+	printf("End Table\n");
+}
 
 void	test_hashtable(void)
 {
